requirements.txt
mcp
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langfuse>=2.0.0
streamlit>=1.35.0
python-dotenv
httpx
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
